Use dict instead of db to get value in processor, pause for exporter.
Now takes on average of ~1 second per update, rather than ~4.

Also increased time for batching to 45-60 from 3-5, which significantly
reduces the number of repeat messages.

Trying to update the database when it's being read by the exporter seems
like a bad idea, so processor is paused when exporter is accessing
database.