Remove exporter lock after getting results, speed up fetching.
Listings exporter only needs the database while retrieving the listings,
so only need to pause the message processor until that's complete.

fetchall() is way too slow, implemented fetchIter() method to speed it
up.

Now takes ~15 seconds to get all the listings, and ~2:45 minutes to
export the file. Not bad for 3 million+ lines.

Message processor is having no trouble at all keeping up with the
listener, but have yet to test how long it takes to catch up to the
backlog caused by a dump update.