Exporter busy check move, start update checker first.
Moved busy signal check to be first thing listings exporter does so that
update checker isn't waiting for the export to finish, which is kind of
pointless anyway.

Made sure both that update checker is started first, and that there is a
delay before the other threads get started, so checker has time to see
if there's an update before the other threads start mucking with the
database, when the program is first launched.