Updated load_config() to enable easily adding additional settings.
load_config() has been redesigned to make it easier to add additional
settings without needing to worry about the state of the configuration
file of upgrading users.

load_config() now begins by assuming the default values for all
settings, then checks for the existence of the file, and the existence
of each setting within the file. Any settings in the file are loaded
into the config, overwriting the default setting.

Once the config has been fully loaded, if the file does not exist, or if
one or more settings were not found in the file, the config is written
to file. Any settings previously in the file will be preserved, any
missing settings will be added with the default value.