Change name so Python stops yelling at me.
Python's been saying
"Invalid name for Python module: eddblink-listener (it'll not be
analyzed)" this whole time. It's bugging me, because that message means
that the program won't be cached, meaning it'll run slower, and it can't
be imported by other programs - if someone decides to make a Python GUI
for this or something.)

So, yeah, now it's an underscore. Sorry for any problems it causes you.