Add validation check for "plugin_options" setting.
Unlink the other settings, if this one is found to be invalid, and valid
values will be retained, and the default values will be prepended to the
setting if they don't already exist.

For example, `"all,force,listings,horseradish"` has the invalid setting
"horseradish", contains the valid value "listings, and is missing the
default value "skipvend", so after validation that setting will be
`"skipvend,all,force,listings"`

If the entirety of the setting is valid, it will not be altered.

For example, `"listings,ship,upgrade,fallback"` has only valid values,
so after validation would still be `"listings,ship,upgrade,fallback"`.