Retry, not crash, if DB locked. Export just listings updated since dump. Check update using listings.csv, not commodities.json
"commodities.json" is the first file to update on EDDB, which means the
update checker is run when only that file has updated, meaning all the
other data is a day old. We don't want that. "listings.csv" is the last
file to be updated, so checking it for an update means the others have
already updated, so we get the latest data.

Listings exporter has been modified to only export the listings that
have been updated since the last EDDB dump, and now exports them to
"listings-live.csv", rather than overwriting the EDDB dump. EDDBlink
plugin must be updated at version 0.26+ or the listener will exit. as
v0.26 is the first version of the plugin that adds the DB data needed
for this feature. ("from_live" column in StationItem table.)

Both "listings.csv" AND "listings-live.csv" are required for a client to
get the most up-to-date data, EDDBlink plugin v0.26+ will attempt to get
and import both.

Listener will now wait for a second and retry SQL execution is attempt
failed due to DB being locked, rather than crashing.