Update dicts when updating, in case any data changed since last.
It is possible that some data has changed since the previous update,
such as new commodities being added, or a mobile station jumping from
one system to another, so we ned to update the dicts to reflect the new
data.

We're also not deleting the system_names dict anymore, because we'll
need that to update mobile stations once I figure out how to do so.