Start listener thread ASAP, handle mobile stations.
Start the listener thread as soon as the bare neccessities have been
loaded, this means missing less messages during the start-up process.

Handle the case where the message is for a mobile station- i.e., one
that can jump to new systems.

Instead of storing their ID# by system/station like with normal
stations, these all have there ID# number stored by MEGASHIP/station.

When the message processor finds a MEGASHIP station, it updates the
system_id for that station in the DB with the current system according
to the message, and continues processing as with normal stations.