Rem env '3.6', del var post check, remove 'U', +"IGNORE", catch errors
For some reason, having the '3.6' part of "#!/usr/bin/env python3.6"
causes my Python 3.7 install to forget it has pyzmq installed and
everything borks.

Delete localModded and dumpModded after checking for update, hopefully
it makes #14 go away.

Remove "U" from the "with open(...)" statements as it is deprecated in
3.7 and will be removed:
"eddblink_listener.py:799: DeprecationWarning: 'U' mode is deprecated
  with open(str(dataPath / Path("Item.csv")), "rU") as fh:"

Added " OR IGNORE" to insStmt, which should cause the executemany not to
fail due to duplicate items in the source message. Should take care of
#15, but just in case:

Put executemany into a try block to catch any error and print the error
and the itemList that caused it to debug log if debug is enabled.
Definitely takes care of #15 and helps in debugging any errors that
occur in future.

(NOTE: "OR IGNORE" should keep the SQL statement from being rolled back
on an IntegrityError by ignoring the offending statement, and continue
working on the rest of itemList, meaning the message still gets
processed, whereas the try block will catch any error that causes the
SQL statement to fail, meaning the message did not get processed, and
prints to debug to help figure out WHY.)