Add server maint. config, update valid plugin_options, change to 'solo'.
Added 'server_maint_every_x_hour' option to config, will run the
following two tasks every 1 <= X <= 24 (default 12) hours as set in the
config:
VACUUM;
PRAGMA optimize;
This only occurs when the listener is being run in server mode, and only
needs to be done when the database is in WAL mode.

Updated the valid plugin_options to include 'progbar' and 'solo' as they
have been added to EDDBlink plugin.

Changed the initial launch of EDDBlink plugin to run with just
'-O solo', as this will cause it to only perform updates needed by the
EDDBlink infrastructure, and will not actually perform any database
updates, which is what the update checker is for.