Update avg_price of items from the EDDN messages.
There are, currently, four salvage items that are not included in
EDDB.io's API, so they have to be added, for now, manually by the
eddblink plugin. Since they aren't in the API, that means the avg_price
doesn't get updated by eddblink updates like the other items do.

To fix that, and as a side benefit keep each item's average price as up
to date as possible, the listener will now update the average price of
each item from the messages it processes.

Going to need to have the listener export a commodities.json file as
well now, though. Next update.