Update listener to use fdev_id to find item_id for an item.
The commodity['name'] is the "symbol" in the source table, so rather
than using the symbol to get the 'name', use it to get the fdev_id, then
use the fdev_id to get the item_id.

If it can't find the item_id for an item, such as when that item hasn't
been added to EDDB.io yet, use the fdev_id as the item_id. EDDBlink is
already set up to detect and update changes to an item's item_id.

(Removed code added for testing purposes.)